requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astroid==2.4.2
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.1
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
flake8==3.8.3
flake8-polyfill==1.0.2
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.3
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.6
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.7.0
invoke==1.4.1
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lizard==1.17.4
mando==0.6.4
Markdown==3.2.1
matplotlib==3.2.2
mccabe==0.6.1
more-itertools==8.4.0
networkx==2.4
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycrypto==2.6.1
PyDriller==1.15.2
pyflakes==2.2.0
Pygments==2.6.1
pygobject==3.26.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
radon==4.1.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.2
scipy==1.4.1
SecretStorage==2.3.1
Shapely==1.7.0
six==1.14.0
smmap==3.0.4
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
tifffile==2020.6.3
toml==0.10.1
tqdm==4.46.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
visual-logging==1.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xenon==0.7.0
zipp==3.1.0